refactor(cli): split tsz entrypoint helpers
