+++ crates/tsz-cli/src/bin/tsz.rs
@@ -3,18 +3,21 @@
 static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 
 use anyhow::{Context, Result};
-use clap::{CommandFactory, Parser};
+use clap::Parser;
 use rustc_hash::FxHashMap;
 use std::ffi::OsString;
 use std::io::IsTerminal;
 use std::path::{Path, PathBuf};
-use std::time::Duration;
 
 use tsz::checker::diagnostics::DiagnosticCategory;
 use tsz_cli::args::CliArgs;
 use tsz_cli::help::{self, TSC_VERSION};
 use tsz_cli::{driver, locale, reporter::Reporter, watch};
 use tsz_common::diagnostics::{diagnostic_codes, diagnostic_messages};
+
+use arg_preprocess::{check_build_position, preprocess_args};
+use clap_errors::handle_clap_error;
+use diagnostics_report::print_diagnostics;
 
 /// tsc exit status codes (matching TypeScript's `ExitStatus` enum)
 const EXIT_SUCCESS: i32 = 0;
@@ -479,1025 +482,6 @@ fn run_batch_mode() -> Result<()> {
     Ok(())
 }
 
-/// Preprocess command-line arguments for tsc compatibility.
-///
-/// Handles (BEFORE clap parsing):
-/// - `--version` / `-v` / `-V` → print version, exit 0
-/// - `--help` / `-h` / `-?` → print help, exit 0
-/// - `--all` (with or without `--help`) → print all options, exit 0
-/// - `@file` response file expansion (tsc reads args from response files)
-/// - Build mode flag remapping: when `--build`/`-b` is the first argument,
-///   `-v` maps to `--build-verbose`, `-d` maps to `--dry`, `-f` maps to `--force`
-/// - Case-insensitive flag names: `--NoEmit` → `--noEmit` (tsc v6 compat)
-/// - Boolean flag values: `--strict false` → strip the flag (tsc v6 compat)
-/// - Optional boolean flags: `--strictNullChecks file.ts` → `--strictNullChecks=true file.ts`
-/// - Duplicate flags: `--strict --strict` → deduplicated (tsc v6 compat)
-fn preprocess_args(args: Vec<OsString>) -> Vec<OsString> {
-    // First pass: expand response files and collect normalized arg strings
-    let mut expanded = Vec::with_capacity(args.len());
-
-    for (i, arg) in args.iter().enumerate() {
-        if i == 0 {
-            expanded.push(arg.clone());
-            continue;
-        }
-
-        let arg_str = arg.to_string_lossy();
-
-        if arg_str.starts_with('@') && arg_str.len() > 1 {
-            // Response file: @path reads arguments from file
-            let path = &arg_str[1..];
-            match std::fs::read_to_string(path) {
-                Ok(content) => {
-                    for line in content.lines() {
-                        let trimmed = line.trim();
-                        if !trimmed.is_empty() {
-                            for part in split_response_line(trimmed) {
-                                expanded.push(OsString::from(part));
-                            }
-                        }
-                    }
-                }
-                Err(_) => {
-                    expanded.push(arg.clone());
-                }
-            }
-        } else {
-            expanded.push(arg.clone());
-        }
-    }
-
-    // Pass 1.5: normalize --Flag names to canonical casing (case-insensitive, tsc compat).
-    // Only the flag name portion is lowercased; values and file paths are preserved as-is.
-    for arg in expanded.iter_mut().skip(1) {
-        let s = arg.to_string_lossy();
-        if s.starts_with("--") && s.len() > 2 {
-            if let Some(eq_pos) = s.find('=') {
-                let flag_part = &s[2..eq_pos];
-                let value_part = &s[eq_pos..];
-                if let Some(canonical) = canonicalize_long_flag(flag_part) {
-                    *arg = OsString::from(format!("{canonical}{value_part}"));
-                }
-            } else {
-                let flag_part = &s[2..];
-                if let Some(canonical) = canonicalize_long_flag(flag_part) {
-                    *arg = OsString::from(canonical);
-                }
-            }
-        }
-    }
-
-    // Second pass: detect --help/-h/-?/--all/--version/-v/-V before clap
-    // In build mode, -v means --build-verbose, not --version
-    let is_build_mode = expanded
-        .get(1)
-        .map(|a| {
-            let s = a.to_string_lossy();
-            s == "--build" || s == "-b"
-        })
-        .unwrap_or(false);
-
-    let mut has_help = false;
-    let mut has_all = false;
-    let mut has_version = false;
-
-    for (i, arg) in expanded.iter().enumerate() {
-        if i == 0 {
-            continue;
-        }
-        let s = arg.to_string_lossy();
-        match s.as_ref() {
-            "--help" | "-h" | "-?" => has_help = true,
-            "--all" => has_all = true,
-            "--version" | "-V" => has_version = true,
-            // -v means version only outside build mode; in build mode it means --build-verbose
-            "-v" if !is_build_mode => has_version = true,
-            _ => {}
-        }
-    }
-
-    // --all takes precedence (with or without --help)
-    if has_all {
-        println!(
-            "{}",
-            help::colorize_help(&help::render_help_all(TSC_VERSION))
-        );
-        std::process::exit(0);
-    }
-
-    // --help / -h / -?
-    if has_help {
-        println!("{}", help::colorize_help(&help::render_help(TSC_VERSION)));
-        std::process::exit(0);
-    }
-
-    // --version / -v / -V
-    if has_version {
-        println!("Version {TSC_VERSION}");
-        std::process::exit(0);
-    }
-
-    // Check for bare `--` and `-` which tsc treats as unknown options (TS5023).
-    // These must be detected before clap sees them (clap treats `--` as
-    // end-of-options and `-` as a positional arg).
-    // Also check for --boolFlag=value which tsc treats as an unknown option.
-    for arg in expanded.iter().skip(1) {
-        let s = arg.to_string_lossy();
-        if s == "--" || s == "-" {
-            println!("error TS5023: Unknown compiler option '{s}'.");
-            std::process::exit(1);
-        }
-        // tsc treats --boolFlag=value as an unknown option (the whole --flag=value string)
-        if let Some(eq_pos) = s.find('=') {
-            let flag_part = &s[..eq_pos];
-            if is_boolean_flag(flag_part) {
-                println!("error TS5023: Unknown compiler option '{s}'.");
-                std::process::exit(1);
-            }
-        }
-    }
-
-    // Third pass: detect build mode and remap flags
-    let is_build_mode = expanded
-        .get(1)
-        .map(|a| {
-            let s = a.to_string_lossy();
-            s == "--build" || s == "-b"
-        })
-        .unwrap_or(false);
-
-    let mut result = Vec::with_capacity(expanded.len());
-    for (i, arg) in expanded.iter().enumerate() {
-        if i == 0 {
-            result.push(arg.clone());
-            continue;
-        }
-
-        let arg_str = arg.to_string_lossy();
-
-        if is_build_mode && i > 1 {
-            // In build mode, remap short flags:
-            //   -v → --build-verbose (not --version)
-            //   -d → --dry           (not --declaration)
-            //   -f → --force
-            match arg_str.as_ref() {
-                "-v" => {
-                    result.push(OsString::from("--build-verbose"));
-                    continue;
-                }
-                "-d" => {
-                    result.push(OsString::from("--dry"));
-                    continue;
-                }
-                "-f" => {
-                    result.push(OsString::from("--force"));
-                    continue;
-                }
-                _ => {}
-            }
-        }
-
-        result.push(arg.clone());
-    }
-
-    // Fourth pass: handle boolean flag values ("true"/"false") and deduplicate flags.
-    // tsc treats `--strict false` as setting strict=false, and accepts duplicate flags silently.
-    let mut final_result = Vec::with_capacity(result.len());
-    let mut flag_positions: FxHashMap<String, usize> = FxHashMap::default();
-    let mut skip_positions: Vec<bool> = Vec::new();
-
-    let mut i = 0;
-    if !result.is_empty() {
-        final_result.push(result[0].clone());
-        skip_positions.push(false);
-        i = 1;
-    }
-
-    while i < result.len() {
-        let arg_str = result[i].to_string_lossy().to_string();
-
-        if arg_str.starts_with("--") {
-            let flag_name = if let Some(eq_pos) = arg_str.find('=') {
-                arg_str[..eq_pos].to_string()
-            } else {
-                arg_str.clone()
-            };
-
-            let is_boolean = is_boolean_flag(flag_name.as_str());
-            let takes_value = is_valued_flag(flag_name.as_str());
-
-            // Check if next arg is "true" or "false" for boolean flags
-            if is_boolean
-                && !arg_str.contains('=')
-                && let Some(next) = result.get(i + 1)
-            {
-                let next_str = next.to_string_lossy();
-                let next_lower = next_str.to_lowercase();
-                if next_lower == "false" {
-                    if is_option_bool_flag(flag_name.as_str()) {
-                        push_option_bool_arg(
-                            &mut final_result,
-                            &mut skip_positions,
-                            &mut flag_positions,
-                            &flag_name,
-                            false,
-                        );
-                        i += 2;
-                        continue;
-                    }
-                    // Plain bool flag: clap can't represent an explicit `false`,
-                    // so strip the `--flag false` pair and forward the intent
-                    // through a hidden side-channel arg. The override pipeline
-                    // reads `args.explicitly_disabled_bool_flags` and uses it to
-                    // flip a `true` value loaded from `tsconfig.json` to `false`.
-                    if let Some(&prev_idx) = flag_positions.get(&flag_name) {
-                        skip_positions[prev_idx] = true;
-                    }
-                    flag_positions.remove(&flag_name);
-                    let bare = flag_name.trim_start_matches("--");
-                    final_result.push(OsString::from(format!(
-                        "--__explicitly-disabled-bool-flag={bare}"
-                    )));
-                    skip_positions.push(false);
-                    i += 2;
-                    continue;
-                } else if next_lower == "true" {
-                    if is_option_bool_flag(flag_name.as_str()) {
-                        push_option_bool_arg(
-                            &mut final_result,
-                            &mut skip_positions,
-                            &mut flag_positions,
-                            &flag_name,
-                            true,
-                        );
-                        i += 2;
-                        continue;
-                    }
-                    // Plain bool flag: keep the flag, skip the "true" token
-                    i += 1;
-                }
-            }
-
-            if is_boolean && !arg_str.contains('=') && is_option_bool_flag(flag_name.as_str()) {
-                push_option_bool_arg(
-                    &mut final_result,
-                    &mut skip_positions,
-                    &mut flag_positions,
-                    &flag_name,
-                    true,
-                );
-                i += 1;
-                continue;
-            }
-
-            // Deduplicate: if we've seen this flag before, mark old position for skip
-            if let Some(&prev_idx) = flag_positions.get(&flag_name) {
-                skip_positions[prev_idx] = true;
-                if takes_value
-                    && !final_result[prev_idx].to_string_lossy().contains('=')
-                    && prev_idx + 1 < skip_positions.len()
-                {
-                    skip_positions[prev_idx + 1] = true;
-                }
-            }
-
-            let current_idx = final_result.len();
-            flag_positions.insert(flag_name, current_idx);
-            final_result.push(OsString::from(&arg_str));
-            skip_positions.push(false);
-
-            if takes_value && !arg_str.contains('=') {
-                i += 1;
-                if i < result.len() {
-                    final_result.push(result[i].clone());
-                    skip_positions.push(false);
-                }
-            }
-        } else {
-            final_result.push(result[i].clone());
-            skip_positions.push(false);
-        }
-
-        i += 1;
-    }
-
-    final_result
-        .into_iter()
-        .zip(skip_positions)
-        .filter_map(|(arg, skip)| if skip { None } else { Some(arg) })
-        .collect()
-}
-
-fn push_option_bool_arg(
-    final_result: &mut Vec<OsString>,
-    skip_positions: &mut Vec<bool>,
-    flag_positions: &mut FxHashMap<String, usize>,
-    flag_name: &str,
-    value: bool,
-) {
-    if let Some(&prev_idx) = flag_positions.get(flag_name) {
-        skip_positions[prev_idx] = true;
-    }
-
-    let current_idx = final_result.len();
-    flag_positions.insert(flag_name.to_string(), current_idx);
-    final_result.push(OsString::from(format!("{flag_name}={value}")));
-    skip_positions.push(false);
-}
-
-/// Return the canonical long flag spelling for tsc-compatible case-insensitive
-/// input, accepting both camelCase and kebab-case spellings.
-fn canonicalize_long_flag(flag: &str) -> Option<&'static str> {
-    for &known in KNOWN_TSC_OPTIONS {
-        if flag_key_matches(&known[2..], flag) {
-            return Some(known);
-        }
-    }
-
-    match normalized_flag_key(flag).as_str() {
-        "buildverbose" | "verbose" => Some("--build-verbose"),
-        "batch" => Some("--batch"),
-        "diagnosticsjson" => Some("--diagnostics-json"),
-        "perfcountersjson" => Some("--perf-counters-json"),
-        "tracedependencies" => Some("--traceDependencies"),
-        "__explicitlydisabledboolflag" => Some("--__explicitly-disabled-bool-flag"),
-        _ => None,
-    }
-}
-
-fn flag_key_matches(canonical: &str, input: &str) -> bool {
-    canonical
-        .bytes()
-        .filter(|&b| b != b'-')
-        .map(|b| b.to_ascii_lowercase())
-        .eq(input
-            .bytes()
-            .filter(|&b| b != b'-')
-            .map(|b| b.to_ascii_lowercase()))
-}
-
-fn normalized_flag_key(flag: &str) -> String {
-    flag.bytes()
-        .filter(|&b| b != b'-')
-        .map(|b| b.to_ascii_lowercase() as char)
-        .collect()
-}
-
-/// Known boolean flags (flags that accept no value or optional true/false).
-const BOOLEAN_FLAGS: &[&str] = &[
-    "--all",
-    "--build",
-    "--init",
-    "--listFilesOnly",
-    "--showConfig",
-    "--ignoreConfig",
-    "--libReplacement",
-    "--watch",
-    "--noLib",
-    "--useDefineForClassFields",
-    "--experimentalDecorators",
-    "--emitDecoratorMetadata",
-    "--resolveJsonModule",
-    "--resolvePackageJsonExports",
-    "--resolvePackageJsonImports",
-    "--allowArbitraryExtensions",
-    "--allowImportingTsExtensions",
-    "--rewriteRelativeImportExtensions",
-    "--noResolve",
-    "--allowUmdGlobalAccess",
-    "--noUncheckedSideEffectImports",
-    "--allowJs",
-    "--checkJs",
-    "--declaration",
-    "--declarationMap",
-    "--emitDeclarationOnly",
-    "--sourceMap",
-    "--inlineSourceMap",
-    "--inlineSources",
-    "--noEmit",
-    "--noEmitOnError",
-    "--noEmitHelpers",
-    "--importHelpers",
-    "--downlevelIteration",
-    "--removeComments",
-    "--preserveConstEnums",
-    "--stripInternal",
-    "--emitBOM",
-    "--esModuleInterop",
-    "--allowSyntheticDefaultImports",
-    "--isolatedModules",
-    "--isolatedDeclarations",
-    "--verbatimModuleSyntax",
-    "--forceConsistentCasingInFileNames",
-    "--preserveSymlinks",
-    "--erasableSyntaxOnly",
-    "--strict",
-    "--noImplicitAny",
-    "--strictNullChecks",
-    "--strictFunctionTypes",
-    "--strictBindCallApply",
-    "--strictPropertyInitialization",
-    "--strictBuiltinIteratorReturn",
-    "--noImplicitThis",
-    "--useUnknownInCatchVariables",
-    "--alwaysStrict",
-    "--noUnusedLocals",
-    "--noUnusedParameters",
-    "--exactOptionalPropertyTypes",
-    "--noImplicitReturns",
-    "--noFallthroughCasesInSwitch",
-    "--sound",
-    "--soundReportOnly",
-    "--noUncheckedIndexedAccess",
-    "--noImplicitOverride",
-    "--noPropertyAccessFromIndexSignature",
-    "--allowUnreachableCode",
-    "--allowUnusedLabels",
-    "--skipDefaultLibCheck",
-    "--skipLibCheck",
-    "--composite",
-    "--incremental",
-    "--disableReferencedProjectLoad",
-    "--disableSolutionSearching",
-    "--disableSourceOfProjectReferenceRedirect",
-    "--diagnostics",
-    "--extendedDiagnostics",
-    "--explainFiles",
-    "--listFiles",
-    "--listEmittedFiles",
-    "--traceResolution",
-    "--traceDependencies",
-    "--noCheck",
-    "--pretty",
-    "--noErrorTruncation",
-    "--preserveWatchOutput",
-    "--synchronousWatchDirectory",
-    "--build-verbose",
-    "--dry",
-    "--force",
-    "--clean",
-    "--stopBuildOnErrors",
-    "--assumeChangesOnlyAffectDirectDependencies",
-    "--keyofStringsOnly",
-    "--noImplicitUseStrict",
-    "--noStrictGenericChecks",
-    "--preserveValueImports",
-    "--suppressExcessPropertyErrors",
-    "--suppressImplicitAnyIndexErrors",
-    "--disableSizeLimit",
-    "--batch",
-];
-
-fn is_boolean_flag(flag: &str) -> bool {
-    BOOLEAN_FLAGS.contains(&flag)
-}
-
-/// Flags that take a mandatory value argument (not boolean flags).
-const VALUED_FLAGS: &[&str] = &[
-    "--locale",
-    "--project",
-    "--target",
-    "--module",
-    "--lib",
-    "--jsx",
-    "--jsxFactory",
-    "--jsxFragmentFactory",
-    "--jsxImportSource",
-    "--moduleDetection",
-    "--moduleResolution",
-    "--baseUrl",
-    "--typeRoots",
-    "--types",
-    "--rootDirs",
-    "--paths",
-    "--plugins",
-    "--moduleSuffixes",
-    "--customConditions",
-    "--maxNodeModuleJsDepth",
-    "--declarationDir",
-    "--outDir",
-    "--rootDir",
-    "--outFile",
-    "--mapRoot",
-    "--sourceRoot",
-    "--newLine",
-    "--tsBuildInfoFile",
-    "--generateTrace",
-    "--generateCpuProfile",
-    "--ignoreDeprecations",
-    "--watchFile",
-    "--watchDirectory",
-    "--fallbackPolling",
-    "--excludeDirectories",
-    "--excludeFiles",
-    "--reactNamespace",
-    "--charset",
-    "--importsNotUsedAsValues",
-    "--out",
-    "--typesVersions",
-];
-
-fn is_valued_flag(flag: &str) -> bool {
-    VALUED_FLAGS.contains(&flag)
-}
-
-/// Flags that are Option<bool> (tri-state: None, Some(true), Some(false)).
-/// These need --flag=true or --flag=false rather than flag removal.
-const OPTION_BOOL_FLAGS: &[&str] = &[
-    "--useDefineForClassFields",
-    "--resolvePackageJsonExports",
-    "--resolvePackageJsonImports",
-    "--allowSyntheticDefaultImports",
-    "--forceConsistentCasingInFileNames",
-    "--noImplicitAny",
-    "--strictNullChecks",
-    "--strictFunctionTypes",
-    "--strictBindCallApply",
-    "--strictPropertyInitialization",
-    "--strictBuiltinIteratorReturn",
-    "--noImplicitThis",
-    "--useUnknownInCatchVariables",
-    "--alwaysStrict",
-    "--allowUnreachableCode",
-    "--allowUnusedLabels",
-    "--pretty",
-];
-
-fn is_option_bool_flag(flag: &str) -> bool {
-    OPTION_BOOL_FLAGS.contains(&flag)
-}
-
-/// Split a response file line into arguments, respecting quoted strings.
-///
-/// Handles both double (`"`) and single (`'`) quotes. Quotes are stripped
-/// from the resulting tokens. Unquoted regions are split on whitespace.
-fn split_response_line(line: &str) -> Vec<String> {
-    let mut args = Vec::new();
-    let mut current = String::new();
-    let mut in_quote: Option<char> = None;
-
-    for ch in line.chars() {
-        match in_quote {
-            Some(q) if ch == q => {
-                // Closing quote — end quoted region but don't push yet,
-                // there may be more content adjacent (e.g. foo"bar"baz)
-                in_quote = None;
-            }
-            Some(_) => {
-                // Inside quotes — take character literally
-                current.push(ch);
-            }
-            None if ch == '"' || ch == '\'' => {
-                // Opening quote
-                in_quote = Some(ch);
-            }
-            None if ch.is_ascii_whitespace() => {
-                // Unquoted whitespace — flush current token
-                if !current.is_empty() {
-                    args.push(std::mem::take(&mut current));
-                }
-            }
-            None => {
-                current.push(ch);
-            }
-        }
-    }
-
-    if !current.is_empty() {
-        args.push(current);
-    }
-
-    args
-}
-
-/// Check that --build/-b is the first argument.
-/// tsc v6 behavior:
-///   - `--build` (long form) not first → TS6369 ("must be first")
-///   - `-b` (short form) not first → TS5023 ("unknown compiler option")
-///
-/// Returns an error message if either form appears but is not first.
-fn check_build_position(args: &[OsString]) -> Option<String> {
-    // Skip program name (index 0)
-    let mut first_non_program = true;
-
-    for arg in args.iter().skip(1) {
-        let s = arg.to_string_lossy();
-        if s == "--build" {
-            if !first_non_program {
-                return Some(
-                    "error TS6369: Option '--build' must be the first command line argument.\n"
-                        .to_string(),
-                );
-            }
-            return None;
-        }
-        if s == "-b" {
-            if !first_non_program {
-                return Some("error TS5023: Unknown compiler option '-b'.\n".to_string());
-            }
-            return None;
-        }
-        first_non_program = false;
-    }
-
-    None
-}
-
-/// Handle a clap parsing error by reformatting it as a tsc-style diagnostic.
-fn handle_clap_error(err: clap::Error, args: &[OsString]) -> Result<()> {
-    use clap::error::ErrorKind;
-
-    match err.kind() {
-        ErrorKind::UnknownArgument => {
-            // Extract ALL unknown flags from the args and report each one
-            let unknown_flags = extract_all_unknown_flags(args);
-            if unknown_flags.is_empty() {
-                // Fallback: just print TS5023 with whatever info we have
-                println!("error TS5023: Unknown compiler option.");
-            } else {
-                for flag in &unknown_flags {
-                    // Try to find a close match for TS5025
-                    if let Some(suggestion) = find_closest_option(flag) {
-                        let suggestion_name = suggestion.strip_prefix("--").unwrap_or(suggestion);
-                        println!(
-                            "error TS5025: Unknown compiler option '{flag}'. Did you mean '{suggestion_name}'?"
-                        );
-                    } else {
-                        println!("error TS5023: Unknown compiler option '{flag}'.");
-                    }
-                }
-            }
-            std::process::exit(1);
-        }
-        ErrorKind::DisplayHelp | ErrorKind::DisplayVersion => {
-            // Help and version are handled in preprocess_args before clap,
-            // but keep this arm for safety
-            err.exit();
-        }
-        ErrorKind::MissingRequiredArgument => {
-            // TS6044: Compiler option 'X' expects an argument.
-            // Extract the option name from clap's error message
-            let msg = err.to_string();
-            if let Some(option_name) = extract_option_from_missing_value(&msg) {
-                println!("error TS6044: Compiler option '{option_name}' expects an argument.");
-                // Also emit TS6046 with valid values if this is an enum option
-                if let Some(valid_values) = get_valid_values_for_option(&option_name) {
-                    println!(
-                        "error TS6046: Argument for '--{option_name}' option must be: {valid_values}."
-                    );
-                }
-            } else {
-                let msg = msg
-                    .lines()
-                    .next()
-                    .unwrap_or(&msg)
-                    .trim_start_matches("error: ");
-                println!("error TS5023: {msg}");
-            }
-            std::process::exit(1);
-        }
-        ErrorKind::InvalidValue => {
-            let msg = err.to_string();
-            // Detect the "missing value" case: clap says "a value is required for"
-            let is_missing_value = msg.contains("a value is required for");
-            if let Some(option_name) = extract_option_from_invalid_value(&msg) {
-                // TS6044: emit when the option was given without any value
-                if is_missing_value {
-                    println!("error TS6044: Compiler option '{option_name}' expects an argument.");
-                }
-                // TS6046: list valid values for enum options
-                if let Some(valid_values) = get_valid_values_for_option(&option_name) {
-                    println!(
-                        "error TS6046: Argument for '--{option_name}' option must be: {valid_values}."
-                    );
-                } else if !is_missing_value {
-                    let msg = msg
-                        .lines()
-                        .next()
-                        .unwrap_or(&msg)
-                        .trim_start_matches("error: ");
-                    println!("error TS5023: {msg}");
-                }
-            } else {
-                let msg = msg
-                    .lines()
-                    .next()
-                    .unwrap_or(&msg)
-                    .trim_start_matches("error: ");
-                println!("error TS5023: {msg}");
-            }
-            std::process::exit(1);
-        }
-        _ => {
-            // For other clap errors, still use exit code 1
-            // and tsc-style formatting where possible
-            let msg = err.to_string();
-            // Strip clap's formatting prefix
-            let msg = msg
-                .lines()
-                .next()
-                .unwrap_or(&msg)
-                .trim_start_matches("error: ");
-            println!("error TS5023: {msg}");
-            std::process::exit(1);
-        }
-    }
-}
-
-/// Extract ALL unknown flags from the preprocessed args.
-/// Scans all args against known options and returns every unrecognized flag.
-fn extract_all_unknown_flags(args: &[OsString]) -> Vec<String> {
-    let known = collect_known_flags();
-    let mut unknown = Vec::new();
-
-    for arg in args.iter().skip(1) {
-        let s = arg.to_string_lossy();
-        if s == "-" || s == "--" {
-            // Bare `-` and `--` are treated as unknown options to match tsc
-            unknown.push(s.into_owned());
-        } else if s.starts_with('-') && !s.starts_with("--") && s.len() == 2 {
-            // Short flag like -x
-            if !known.iter().any(|k| k == s.as_ref()) {
-                unknown.push(s.into_owned());
-            }
-        } else if s.starts_with("--") {
-            // Long flag like --badFlag (may have =value)
-            let flag_part = s.split('=').next().unwrap_or(&s);
-            if !known.iter().any(|k| k == flag_part) {
-                unknown.push(flag_part.to_string());
-            }
-        }
-    }
-    unknown
-}
-
-/// Collect all known CLI flags from clap's command definition.
-fn collect_known_flags() -> Vec<String> {
-    let cmd = CliArgs::command();
-    let mut known: Vec<String> = Vec::new();
-    for a in cmd.get_arguments() {
-        if let Some(long) = a.get_long() {
-            known.push(format!("--{long}"));
-        }
-        // get_visible_aliases returns individual aliases
-        for alias in a.get_visible_aliases().unwrap_or_default() {
-            known.push(format!("--{alias}"));
-        }
-        if let Some(short) = a.get_short() {
-            known.push(format!("-{short}"));
-        }
-    }
-    // Also add -V (clap's version flag after our -v remapping)
-    known.push("-V".to_string());
-    known.push("--version".to_string());
-    known.push("--help".to_string());
-    known.push("-h".to_string());
-    known
-}
-
-/// Extract the option name from a clap "missing required argument" error.
-/// Clap formats these as: "a value is required for '--target <TARGET>' but none was supplied"
-fn extract_option_from_missing_value(msg: &str) -> Option<String> {
-    // Look for pattern: '--optionName' or '--optionName <VALUE>'
-    let start = msg.find("'--")?;
-    let after = &msg[start + 3..];
-    let end = after.find(['\'', ' ', '<'])?;
-    Some(after[..end].to_string())
-}
-
-/// Extract the option name from a clap "invalid value" error.
-/// Clap formats these as: "invalid value 'blah' for '--target <TARGET>'"
-fn extract_option_from_invalid_value(msg: &str) -> Option<String> {
-    let start = msg.find("'--")?;
-    let after = &msg[start + 3..];
-    let end = after.find(['\'', ' ', '<'])?;
-    Some(after[..end].to_string())
-}
-
-/// Get the valid values string for enum-typed CLI options, matching tsc's TS6046 format.
-fn get_valid_values_for_option(option_name: &str) -> Option<&'static str> {
-    // Value ordering and inclusion matches tsc baselines exactly.
-    match option_name {
-        "target" => Some(
-            "'es6', 'es2015', 'es2016', 'es2017', 'es2018', 'es2019', 'es2020', 'es2021', 'es2022', 'es2023', 'es2024', 'es2025', 'esnext'",
-        ),
-        "module" => Some(
-            "'commonjs', 'es6', 'es2015', 'es2020', 'es2022', 'esnext', 'node16', 'node18', 'node20', 'nodenext', 'preserve'",
-        ),
-        "jsx" => Some("'preserve', 'react-native', 'react-jsx', 'react-jsxdev', 'react'"),
-        "moduleResolution" | "module-resolution" | "moduleresolution" => {
-            Some("'node16', 'nodenext', 'bundler'")
-        }
-        "moduleDetection" | "module-detection" | "moduledetection" => {
-            Some("'auto', 'legacy', 'force'")
-        }
-        _ => None,
-    }
-}
-
-/// All known tsc compiler option long names (for edit-distance matching).
-/// These are the canonical --camelCase forms that tsc recognizes.
-const KNOWN_TSC_OPTIONS: &[&str] = &[
-    "--all",
-    "--allowArbitraryExtensions",
-    "--allowImportingTsExtensions",
-    "--allowJs",
-    "--allowSyntheticDefaultImports",
-    "--allowUmdGlobalAccess",
-    "--allowUnreachableCode",
-    "--allowUnusedLabels",
-    "--alwaysStrict",
-    "--assumeChangesOnlyAffectDirectDependencies",
-    "--baseUrl",
-    "--build",
-    "--charset",
-    "--checkJs",
-    "--clean",
-    "--composite",
-    "--customConditions",
-    "--declaration",
-    "--declarationDir",
-    "--declarationMap",
-    "--diagnostics",
-    "--disableReferencedProjectLoad",
-    "--disableSizeLimit",
-    "--disableSolutionSearching",
-    "--disableSourceOfProjectReferenceRedirect",
-    "--downlevelIteration",
-    "--dry",
-    "--emitBOM",
-    "--emitDeclarationOnly",
-    "--emitDecoratorMetadata",
-    "--erasableSyntaxOnly",
-    "--esModuleInterop",
-    "--exactOptionalPropertyTypes",
-    "--excludeDirectories",
-    "--excludeFiles",
-    "--experimentalDecorators",
-    "--explainFiles",
-    "--extendedDiagnostics",
-    "--fallbackPolling",
-    "--force",
-    "--forceConsistentCasingInFileNames",
-    "--generateCpuProfile",
-    "--generateTrace",
-    "--help",
-    "--ignoreConfig",
-    "--ignoreDeprecations",
-    "--importHelpers",
-    "--importsNotUsedAsValues",
-    "--incremental",
-    "--init",
-    "--inlineSourceMap",
-    "--inlineSources",
-    "--isolatedDeclarations",
-    "--isolatedModules",
-    "--jsx",
-    "--jsxFactory",
-    "--jsxFragmentFactory",
-    "--jsxImportSource",
-    "--keyofStringsOnly",
-    "--lib",
-    "--libReplacement",
-    "--listEmittedFiles",
-    "--listFiles",
-    "--listFilesOnly",
-    "--locale",
-    "--mapRoot",
-    "--maxNodeModuleJsDepth",
-    "--module",
-    "--moduleDetection",
-    "--moduleResolution",
-    "--moduleSuffixes",
-    "--newLine",
-    "--noCheck",
-    "--noEmit",
-    "--noEmitHelpers",
-    "--noEmitOnError",
-    "--noErrorTruncation",
-    "--noFallthroughCasesInSwitch",
-    "--noImplicitAny",
-    "--noImplicitOverride",
-    "--noImplicitReturns",
-    "--noImplicitThis",
-    "--noImplicitUseStrict",
-    "--noLib",
-    "--noPropertyAccessFromIndexSignature",
-    "--noResolve",
-    "--noStrictGenericChecks",
-    "--noUncheckedIndexedAccess",
-    "--noUncheckedSideEffectImports",
-    "--noUnusedLocals",
-    "--noUnusedParameters",
-    "--out",
-    "--outDir",
-    "--outFile",
-    "--paths",
-    "--plugins",
-    "--preserveConstEnums",
-    "--preserveSymlinks",
-    "--preserveValueImports",
-    "--preserveWatchOutput",
-    "--pretty",
-    "--project",
-    "--reactNamespace",
-    "--removeComments",
-    "--resolveJsonModule",
-    "--resolvePackageJsonExports",
-    "--resolvePackageJsonImports",
-    "--rewriteRelativeImportExtensions",
-    "--rootDir",
-    "--rootDirs",
-    "--showConfig",
-    "--skipDefaultLibCheck",
-    "--skipLibCheck",
-    "--sound",
-    "--soundReportOnly",
-    "--sourceMap",
-    "--sourceRoot",
-    "--stopBuildOnErrors",
-    "--strict",
-    "--strictBindCallApply",
-    "--strictBuiltinIteratorReturn",
-    "--strictFunctionTypes",
-    "--strictNullChecks",
-    "--strictPropertyInitialization",
-    "--strict",
-    "--stripInternal",
-    "--suppressExcessPropertyErrors",
-    "--suppressImplicitAnyIndexErrors",
-    "--synchronousWatchDirectory",
-    "--target",
-    "--traceResolution",
-    "--tsBuildInfoFile",
-    "--typeRoots",
-    "--types",
-    "--typesVersions",
-    "--useDefineForClassFields",
-    "--useUnknownInCatchVariables",
-    "--verbatimModuleSyntax",
-    "--version",
-    "--watch",
-    "--watchDirectory",
-    "--watchFile",
-];
-
-/// Compute Levenshtein edit distance between two strings (case-insensitive).
-fn edit_distance(a: &str, b: &str) -> usize {
-    let a_lower = a.to_lowercase();
-    let b_lower = b.to_lowercase();
-    let a_chars: Vec<char> = a_lower.chars().collect();
-    let b_chars: Vec<char> = b_lower.chars().collect();
-    let m = a_chars.len();
-    let n = b_chars.len();
-
-    let mut dp = vec![vec![0usize; n + 1]; m + 1];
-    for (i, row) in dp.iter_mut().enumerate().take(m + 1) {
-        row[0] = i;
-    }
-    for (j, val) in dp[0].iter_mut().enumerate().take(n + 1) {
-        *val = j;
-    }
-    for i in 1..=m {
-        for j in 1..=n {
-            let cost = if a_chars[i - 1] == b_chars[j - 1] {
-                0
-            } else {
-                1
-            };
-            dp[i][j] = (dp[i - 1][j] + 1)
-                .min(dp[i][j - 1] + 1)
-                .min(dp[i - 1][j - 1] + cost);
-        }
-    }
-    dp[m][n]
-}
-
-/// Find the closest known tsc option to the given unknown flag.
-/// Returns `Some(suggestion)` if a reasonably close match exists (edit distance <= 3).
-fn find_closest_option(unknown: &str) -> Option<&'static str> {
-    let mut best: Option<(&str, usize)> = None;
-    for &known in KNOWN_TSC_OPTIONS {
-        let dist = edit_distance(unknown, known);
-        if let Some((_, best_dist)) = best {
-            if dist < best_dist {
-                best = Some((known, dist));
-            }
-        } else {
-            best = Some((known, dist));
-        }
-    }
-
-    // Only suggest if the distance is small enough to be a plausible typo.
-    // tsc uses a threshold proportional to the option name length.
-    // We use max(unknown_len, candidate_len) * 0.4 as the cutoff, with a minimum of 1.
-    best.and_then(|(name, dist)| {
-        let max_len = unknown.len().max(name.len());
-        let threshold = (max_len * 2 / 5).max(1); // ~40% of the longer name
-        if dist <= threshold { Some(name) } else { None }
-    })
-}
-
 /// Read the discovered tsconfig and OR its output-only `compilerOptions`
 /// flags into `args`. tsc honors `listFiles`, `listEmittedFiles`,
 /// `explainFiles`, `diagnostics`, `extendedDiagnostics`, and
@@ -1563,742 +547,6 @@ fn merge_output_only_options_from_tsconfig(args: &mut CliArgs, cwd: &std::path::
     take_bool("traceResolution", &mut args.trace_resolution);
 }
 
-/// Line counts categorized by source-file type, matching tsc's `--diagnostics` output.
-#[derive(Debug, Default, PartialEq, Eq)]
-struct FileLinesStats {
-    library: u64,
-    definitions: u64,
-    typescript: u64,
-    javascript: u64,
-    json: u64,
-    other: u64,
-}
-
-/// Read each file and count its lines, grouped by source-file category.
-///
-/// This is the only I/O-performing step in the diagnostics pipeline; all
-/// downstream rendering operates on the returned counts.
-fn collect_file_lines(files: &[PathBuf]) -> FileLinesStats {
-    let mut stats = FileLinesStats::default();
-    for path in files {
-        let count = std::fs::read_to_string(path)
-            .ok()
-            .map_or(0, |text| text.lines().count() as u64);
-        let name = path.to_string_lossy();
-        if name.contains("lib.") && name.ends_with(".d.ts") {
-            stats.library += count;
-        } else if name.ends_with(".d.ts") || name.ends_with(".d.mts") || name.ends_with(".d.cts") {
-            stats.definitions += count;
-        } else if name.ends_with(".ts")
-            || name.ends_with(".tsx")
-            || name.ends_with(".mts")
-            || name.ends_with(".cts")
-        {
-            stats.typescript += count;
-        } else if name.ends_with(".js")
-            || name.ends_with(".jsx")
-            || name.ends_with(".mjs")
-            || name.ends_with(".cjs")
-        {
-            stats.javascript += count;
-        } else if name.ends_with(".json") {
-            stats.json += count;
-        } else {
-            stats.other += count;
-        }
-    }
-    stats
-}
-
-/// Flat snapshot of all values needed to render the `--diagnostics` /
-/// `--extendedDiagnostics` report.  Contains only primitive types so it is
-/// easy to construct in unit tests without pulling in solver/checker types.
-#[derive(Debug, Default)]
-struct DiagnosticsReport {
-    // Basic (always rendered)
-    files_count: usize,
-    lines: FileLinesStats,
-    error_count: usize,
-    has_phase_timings: bool,
-    io_read_secs: f64,
-    parse_bind_secs: f64,
-    check_secs: f64,
-    emit_secs: f64,
-    total_secs: f64,
-    // Extended section
-    memory_used_kb: u64,
-    emitted_files_count: usize,
-    total_diagnostics: usize,
-    request_cache_hits: u64,
-    request_cache_misses: u64,
-    contextual_cache_bypasses: u64,
-    clear_type_cache_recursive_calls: u64,
-    property_access_cache_hits: u64,
-    property_access_cache_lookups: u64,
-    interned_types_count: usize,
-    interner_kb: f64,
-    // Solver relation caches (present when query_cache_stats is Some)
-    has_query_cache: bool,
-    subtype_entries: usize,
-    subtype_hits: u64,
-    subtype_misses: u64,
-    assignability_entries: usize,
-    assignability_hits: u64,
-    assignability_misses: u64,
-    eval_cache_entries: usize,
-    property_cache_entries: usize,
-    variance_cache_entries: usize,
-    query_cache_kb: f64,
-    // Definition store (present when def_store_stats is Some)
-    has_def_store: bool,
-    def_total: usize,
-    def_type_aliases: usize,
-    def_interfaces: usize,
-    def_classes: usize,
-    def_enums: usize,
-    def_type_to_def_entries: usize,
-    def_symbol_def_index_entries: usize,
-    def_body_to_alias_entries: usize,
-    def_shape_to_def_entries: usize,
-    def_store_kb: f64,
-    // AST residency (present when residency_stats is Some)
-    has_residency: bool,
-    residency_unique_arena_count: usize,
-    residency_arena_kb: f64,
-    residency_file_count: usize,
-    residency_bound_kb: f64,
-    residency_has_pre_merge: bool,
-    residency_pre_merge_kb: f64,
-    residency_has_skeleton: bool,
-    residency_skeleton_symbol_count: usize,
-    residency_skeleton_merge_candidate_count: usize,
-    residency_skeleton_kb: f64,
-    // Module dependency graph (present when module_dep_stats is Some)
-    has_module_deps: bool,
-    module_file_count: usize,
-    module_dependency_edges: usize,
-    module_import_cycles: usize,
-    module_largest_cycle: usize,
-    // Perf-counter dump (non-empty only when TSZ_PERF_COUNTERS is set)
-    perf_counter_dump: String,
-}
-
-/// Collect all data needed for the diagnostics report from a `CompilationResult`.
-///
-/// Performs no I/O; the caller is responsible for supplying `file_lines`
-/// (from `collect_file_lines`) and `memory_used_kb` (from `get_memory_usage_kb`).
-fn build_diagnostics_report(
-    result: &driver::CompilationResult,
-    file_lines: FileLinesStats,
-    elapsed: Duration,
-    memory_used_kb: u64,
-    extended: bool,
-) -> DiagnosticsReport {
-    let pt = &result.phase_timings;
-    let error_count = result
-        .diagnostics
-        .iter()
-        .filter(|d| d.category == DiagnosticCategory::Error)
-        .count();
-
-    let mut report = DiagnosticsReport {
-        files_count: result.files_read.len(),
-        lines: file_lines,
-        error_count,
-        has_phase_timings: pt.total_ms > 0.0,
-        io_read_secs: pt.io_read_ms / 1000.0,
-        parse_bind_secs: (pt.load_libs_ms + pt.parse_bind_ms) / 1000.0,
-        check_secs: pt.check_ms / 1000.0,
-        emit_secs: pt.emit_ms / 1000.0,
-        total_secs: elapsed.as_secs_f64(),
-        ..DiagnosticsReport::default()
-    };
-
-    if !extended {
-        return report;
-    }
-
-    let counters = result.request_cache_counters;
-    report.memory_used_kb = memory_used_kb;
-    report.emitted_files_count = result.emitted_files.len();
-    report.total_diagnostics = result.diagnostics.len();
-    report.request_cache_hits = counters.request_cache_hits;
-    report.request_cache_misses = counters.request_cache_misses;
-    report.contextual_cache_bypasses = counters.contextual_cache_bypasses;
-    report.clear_type_cache_recursive_calls = counters.clear_type_cache_recursive_calls;
-    report.property_access_cache_hits = counters.property_access_request_cache_hits;
-    report.property_access_cache_lookups = counters.property_access_request_cache_lookups;
-    report.interned_types_count = result.interned_types_count;
-    report.interner_kb = result.interner_estimated_bytes as f64 / 1024.0;
-
-    if let Some(ref qc) = result.query_cache_stats {
-        report.has_query_cache = true;
-        report.subtype_entries = qc.relation.subtype_entries;
-        report.subtype_hits = qc.relation.subtype_hits;
-        report.subtype_misses = qc.relation.subtype_misses;
-        report.assignability_entries = qc.relation.assignability_entries;
-        report.assignability_hits = qc.relation.assignability_hits;
-        report.assignability_misses = qc.relation.assignability_misses;
-        report.eval_cache_entries = qc.eval_cache_entries;
-        report.property_cache_entries = qc.property_cache_entries;
-        report.variance_cache_entries = qc.variance_cache_entries;
-        report.query_cache_kb = qc.estimated_size_bytes() as f64 / 1024.0;
-    }
-
-    if let Some(ref ds) = result.def_store_stats {
-        report.has_def_store = true;
-        report.def_total = ds.total_definitions;
-        report.def_type_aliases = ds.type_aliases;
-        report.def_interfaces = ds.interfaces;
-        report.def_classes = ds.classes;
-        report.def_enums = ds.enums;
-        report.def_type_to_def_entries = ds.type_to_def_entries;
-        report.def_symbol_def_index_entries = ds.symbol_def_index_entries;
-        report.def_body_to_alias_entries = ds.body_to_alias_entries;
-        report.def_shape_to_def_entries = ds.shape_to_def_entries;
-        report.def_store_kb = ds.estimated_size_bytes as f64 / 1024.0;
-    }
-
-    if let Some(ref rs) = result.residency_stats {
-        report.has_residency = true;
-        report.residency_unique_arena_count = rs.unique_arena_count;
-        report.residency_arena_kb = rs.unique_arena_estimated_bytes as f64 / 1024.0;
-        report.residency_file_count = rs.file_count;
-        report.residency_bound_kb = rs.total_bound_file_bytes as f64 / 1024.0;
-        report.residency_has_pre_merge = rs.pre_merge_bind_total_bytes > 0;
-        report.residency_pre_merge_kb = rs.pre_merge_bind_total_bytes as f64 / 1024.0;
-        report.residency_has_skeleton = rs.has_skeleton_index;
-        report.residency_skeleton_symbol_count = rs.skeleton_total_symbol_count;
-        report.residency_skeleton_merge_candidate_count = rs.skeleton_merge_candidate_count;
-        report.residency_skeleton_kb = rs.skeleton_estimated_size_bytes as f64 / 1024.0;
-    }
-
-    if let Some(ref md) = result.module_dep_stats {
-        report.has_module_deps = true;
-        report.module_file_count = md.file_count;
-        report.module_dependency_edges = md.dependency_edges;
-        report.module_import_cycles = md.import_cycles;
-        report.module_largest_cycle = md.largest_cycle_size;
-    }
-
-    report.perf_counter_dump = tsz_common::perf_counters::PerfCounters::dump_string();
-
-    report
-}
-
-/// Render a `--diagnostics` / `--extendedDiagnostics` report as a `String`.
-///
-/// This function is pure: it reads only from `report` and writes only to the
-/// returned `String`.  All I/O (file reads, memory probing, env-var access)
-/// must have been completed by the caller before calling this function.
-fn render_diagnostics_report(report: &DiagnosticsReport, extended: bool) -> String {
-    use std::fmt::Write as _;
-    let mut out = String::new();
-
-    let _ = writeln!(out);
-    let _ = writeln!(out, "Files:                         {}", report.files_count);
-    let _ = writeln!(
-        out,
-        "Lines of Library:              {}",
-        report.lines.library
-    );
-    let _ = writeln!(
-        out,
-        "Lines of Definitions:          {}",
-        report.lines.definitions
-    );
-    let _ = writeln!(
-        out,
-        "Lines of TypeScript:           {}",
-        report.lines.typescript
-    );
-    let _ = writeln!(
-        out,
-        "Lines of JavaScript:           {}",
-        report.lines.javascript
-    );
-    let _ = writeln!(out, "Lines of JSON:                 {}", report.lines.json);
-    let _ = writeln!(out, "Lines of Other:                {}", report.lines.other);
-    let _ = writeln!(out, "Errors:                        {}", report.error_count);
-
-    if report.has_phase_timings {
-        let _ = writeln!(
-            out,
-            "I/O Read:                      {:.2}s",
-            report.io_read_secs
-        );
-        let _ = writeln!(
-            out,
-            "Parse & Bind:                  {:.2}s",
-            report.parse_bind_secs
-        );
-        let _ = writeln!(
-            out,
-            "Check:                         {:.2}s",
-            report.check_secs
-        );
-        let _ = writeln!(
-            out,
-            "Emit:                          {:.2}s",
-            report.emit_secs
-        );
-    }
-    let _ = writeln!(
-        out,
-        "Total time:                    {:.2}s",
-        report.total_secs
-    );
-
-    if !extended {
-        return out;
-    }
-
-    let _ = writeln!(
-        out,
-        "Emitted files:                 {}",
-        report.emitted_files_count
-    );
-    let _ = writeln!(
-        out,
-        "Total diagnostics:             {}",
-        report.total_diagnostics
-    );
-    let _ = writeln!(
-        out,
-        "Request cache hits:            {}",
-        report.request_cache_hits
-    );
-    let _ = writeln!(
-        out,
-        "Request cache misses:          {}",
-        report.request_cache_misses
-    );
-
-    let request_lookups = report.request_cache_hits + report.request_cache_misses;
-    let request_hit_rate = if request_lookups == 0 {
-        0.0
-    } else {
-        report.request_cache_hits as f64 * 100.0 / request_lookups as f64
-    };
-    let _ = writeln!(out, "Request cache hit rate:        {request_hit_rate:.1}%");
-    let _ = writeln!(
-        out,
-        "Contextual cache bypasses:     {}",
-        report.contextual_cache_bypasses
-    );
-    let _ = writeln!(
-        out,
-        "clear_type_cache_recursive:    {}",
-        report.clear_type_cache_recursive_calls
-    );
-
-    let access_hit_rate = if report.property_access_cache_lookups == 0 {
-        0.0
-    } else {
-        report.property_access_cache_hits as f64 * 100.0
-            / report.property_access_cache_lookups as f64
-    };
-    let _ = writeln!(
-        out,
-        "Access request-cache hit rate: {:.1}% ({}/{})",
-        access_hit_rate, report.property_access_cache_hits, report.property_access_cache_lookups
-    );
-
-    if report.interned_types_count > 0 {
-        let _ = writeln!(
-            out,
-            "Interned types:                {}",
-            report.interned_types_count
-        );
-    }
-
-    if report.has_query_cache {
-        let sub_total = report.subtype_hits + report.subtype_misses;
-        let sub_rate = if sub_total == 0 {
-            0.0
-        } else {
-            report.subtype_hits as f64 * 100.0 / sub_total as f64
-        };
-        let assign_total = report.assignability_hits + report.assignability_misses;
-        let assign_rate = if assign_total == 0 {
-            0.0
-        } else {
-            report.assignability_hits as f64 * 100.0 / assign_total as f64
-        };
-        let _ = writeln!(
-            out,
-            "Subtype cache:                 {} entries ({} hits, {} misses, {sub_rate:.1}%)",
-            report.subtype_entries, report.subtype_hits, report.subtype_misses,
-        );
-        let _ = writeln!(
-            out,
-            "Assignability cache:           {} entries ({} hits, {} misses, {assign_rate:.1}%)",
-            report.assignability_entries, report.assignability_hits, report.assignability_misses,
-        );
-        let _ = writeln!(
-            out,
-            "Eval cache:                    {}",
-            report.eval_cache_entries
-        );
-        let _ = writeln!(
-            out,
-            "Property cache:                {}",
-            report.property_cache_entries
-        );
-        let _ = writeln!(
-            out,
-            "Variance cache:                {}",
-            report.variance_cache_entries
-        );
-    }
-
-    if report.has_def_store {
-        let _ = writeln!(
-            out,
-            "Definitions:                   {} total ({} aliases, {} interfaces, {} classes, {} enums)",
-            report.def_total,
-            report.def_type_aliases,
-            report.def_interfaces,
-            report.def_classes,
-            report.def_enums,
-        );
-        let _ = writeln!(
-            out,
-            "Def indices:                   type_to_def={}, symbol_def={}, body_to_alias={}, shape_to_def={}",
-            report.def_type_to_def_entries,
-            report.def_symbol_def_index_entries,
-            report.def_body_to_alias_entries,
-            report.def_shape_to_def_entries,
-        );
-    }
-
-    let solver_total_kb = report.interner_kb + report.query_cache_kb + report.def_store_kb;
-    if solver_total_kb > 0.0 {
-        let _ = writeln!(
-            out,
-            "Type interner memory:          {:.1}K ({} types)",
-            report.interner_kb, report.interned_types_count,
-        );
-        let _ = writeln!(
-            out,
-            "Query cache memory:            {:.1}K",
-            report.query_cache_kb
-        );
-        let _ = writeln!(
-            out,
-            "Definition store memory:       {:.1}K",
-            report.def_store_kb
-        );
-        let _ = writeln!(out, "Solver total memory:           {solver_total_kb:.1}K");
-    }
-
-    if report.has_residency {
-        let _ = writeln!(
-            out,
-            "AST arenas:                    {} unique ({:.1}K)",
-            report.residency_unique_arena_count, report.residency_arena_kb,
-        );
-        let _ = writeln!(
-            out,
-            "Bound files:                   {} ({:.1}K)",
-            report.residency_file_count, report.residency_bound_kb,
-        );
-        if report.residency_has_pre_merge {
-            let _ = writeln!(
-                out,
-                "Pre-merge bind data:           {:.1}K",
-                report.residency_pre_merge_kb
-            );
-        }
-        if report.residency_has_skeleton {
-            let _ = writeln!(
-                out,
-                "Skeleton index:                {} symbols, {} merge candidates ({:.1}K)",
-                report.residency_skeleton_symbol_count,
-                report.residency_skeleton_merge_candidate_count,
-                report.residency_skeleton_kb,
-            );
-        }
-    }
-
-    if report.has_module_deps {
-        let _ = writeln!(
-            out,
-            "Module files:                  {}",
-            report.module_file_count
-        );
-        let _ = writeln!(
-            out,
-            "Dependency edges:              {}",
-            report.module_dependency_edges
-        );
-        let _ = writeln!(
-            out,
-            "Import cycles:                 {}",
-            report.module_import_cycles
-        );
-        if report.module_largest_cycle > 0 {
-            let _ = writeln!(
-                out,
-                "Largest cycle:                 {} files",
-                report.module_largest_cycle
-            );
-        }
-    }
-
-    if report.memory_used_kb > 0 {
-        let _ = writeln!(
-            out,
-            "Memory used:                   {}K",
-            report.memory_used_kb
-        );
-    }
-
-    if !report.perf_counter_dump.is_empty() {
-        out.push_str(&report.perf_counter_dump);
-    }
-
-    out
-}
-
-/// Thin orchestrator: collect data, render, then emit to stdout.
-fn print_diagnostics(result: &driver::CompilationResult, elapsed: Duration, extended: bool) {
-    let file_lines = collect_file_lines(&result.files_read);
-    let memory_used_kb = if extended { get_memory_usage_kb() } else { 0 };
-    let report = build_diagnostics_report(result, file_lines, elapsed, memory_used_kb, extended);
-    print!("{}", render_diagnostics_report(&report, extended));
-}
-
-#[cfg(test)]
-mod diagnostics_report_tests {
-    use super::*;
-
-    fn basic_report() -> DiagnosticsReport {
-        DiagnosticsReport {
-            files_count: 3,
-            lines: FileLinesStats {
-                library: 100,
-                definitions: 50,
-                typescript: 200,
-                ..FileLinesStats::default()
-            },
-            error_count: 2,
-            total_secs: 1.23,
-            ..DiagnosticsReport::default()
-        }
-    }
-
-    #[test]
-    fn basic_output_contains_required_fields() {
-        let report = basic_report();
-        let out = render_diagnostics_report(&report, false);
-        assert!(
-            out.contains("Files:                         3"),
-            "files count"
-        );
-        assert!(
-            out.contains("Lines of Library:              100"),
-            "library lines"
-        );
-        assert!(
-            out.contains("Lines of TypeScript:           200"),
-            "typescript lines"
-        );
-        assert!(out.contains("Errors:                        2"), "errors");
-        assert!(out.contains("Total time:                    1.23s"), "time");
-    }
-
-    #[test]
-    fn basic_output_excludes_extended_fields() {
-        let report = basic_report();
-        let out = render_diagnostics_report(&report, false);
-        assert!(
-            !out.contains("Request cache"),
-            "no cache stats in basic mode"
-        );
-        assert!(!out.contains("Memory used"), "no memory in basic mode");
-    }
-
-    #[test]
-    fn extended_output_includes_cache_stats() {
-        let report = DiagnosticsReport {
-            files_count: 1,
-            total_secs: 0.5,
-            request_cache_hits: 80,
-            request_cache_misses: 20,
-            has_query_cache: true,
-            subtype_entries: 10,
-            subtype_hits: 8,
-            subtype_misses: 2,
-            assignability_entries: 5,
-            assignability_hits: 3,
-            assignability_misses: 2,
-            memory_used_kb: 4096,
-            ..DiagnosticsReport::default()
-        };
-        let out = render_diagnostics_report(&report, true);
-        assert!(
-            out.contains("Request cache hit rate:        80.0%"),
-            "hit rate: {out}"
-        );
-        assert!(out.contains("Subtype cache:"), "subtype cache");
-        assert!(
-            out.contains("Memory used:                   4096K"),
-            "memory"
-        );
-    }
-
-    #[test]
-    fn phase_timings_only_shown_when_present() {
-        let without = DiagnosticsReport {
-            has_phase_timings: false,
-            total_secs: 0.1,
-            ..DiagnosticsReport::default()
-        };
-        let with_timings = DiagnosticsReport {
-            has_phase_timings: true,
-            io_read_secs: 0.05,
-            parse_bind_secs: 0.03,
-            check_secs: 0.02,
-            emit_secs: 0.01,
-            total_secs: 0.1,
-            ..DiagnosticsReport::default()
-        };
-        let out_without = render_diagnostics_report(&without, false);
-        let out_with = render_diagnostics_report(&with_timings, false);
-        assert!(
-            !out_without.contains("I/O Read:"),
-            "no timings without flag"
-        );
-        assert!(
-            out_with.contains("I/O Read:                      0.05s"),
-            "has timings"
-        );
-    }
-
-    #[test]
-    fn collect_file_lines_categorizes_correctly() {
-        // Build a temp dir with files of different types to verify categorization.
-        let dir = std::env::temp_dir().join("tsz_test_file_lines");
-        let _ = std::fs::create_dir_all(&dir);
-
-        let lib_d_ts = dir.join("lib.es5.d.ts");
-        let user_ts = dir.join("user.ts");
-        let js_file = dir.join("helper.js");
-
-        std::fs::write(&lib_d_ts, "line1\nline2\nline3\n").unwrap();
-        std::fs::write(&user_ts, "line1\nline2\n").unwrap();
-        std::fs::write(&js_file, "line1\n").unwrap();
-
-        let stats = collect_file_lines(&[lib_d_ts, user_ts, js_file]);
-
-        assert_eq!(stats.library, 3, "lib.d.ts lines");
-        assert_eq!(stats.typescript, 2, "ts lines");
-        assert_eq!(stats.javascript, 1, "js lines");
-        assert_eq!(stats.definitions, 0);
-        assert_eq!(stats.json, 0);
-        assert_eq!(stats.other, 0);
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn basic_golden_full_output() {
-        let report = DiagnosticsReport {
-            files_count: 5,
-            lines: FileLinesStats {
-                library: 1000,
-                definitions: 200,
-                typescript: 300,
-                javascript: 50,
-                json: 10,
-                other: 0,
-            },
-            error_count: 3,
-            has_phase_timings: true,
-            io_read_secs: 0.10,
-            parse_bind_secs: 0.20,
-            check_secs: 0.30,
-            emit_secs: 0.05,
-            total_secs: 0.65,
-            ..DiagnosticsReport::default()
-        };
-        let out = render_diagnostics_report(&report, false);
-        let expected = "\n\
-Files:                         5\n\
-Lines of Library:              1000\n\
-Lines of Definitions:          200\n\
-Lines of TypeScript:           300\n\
-Lines of JavaScript:           50\n\
-Lines of JSON:                 10\n\
-Lines of Other:                0\n\
-Errors:                        3\n\
-I/O Read:                      0.10s\n\
-Parse & Bind:                  0.20s\n\
-Check:                         0.30s\n\
-Emit:                          0.05s\n\
-Total time:                    0.65s\n";
-        assert_eq!(out, expected, "basic golden mismatch:\n{out}");
-    }
-
-    #[test]
-    fn extended_golden_full_output() {
-        let report = DiagnosticsReport {
-            files_count: 2,
-            lines: FileLinesStats {
-                library: 500,
-                definitions: 0,
-                typescript: 100,
-                javascript: 0,
-                json: 0,
-                other: 0,
-            },
-            error_count: 0,
-            has_phase_timings: false,
-            total_secs: 1.00,
-            // Extended fields
-            memory_used_kb: 8192,
-            emitted_files_count: 1,
-            total_diagnostics: 0,
-            request_cache_hits: 90,
-            request_cache_misses: 10,
-            contextual_cache_bypasses: 2,
-            clear_type_cache_recursive_calls: 1,
-            property_access_cache_hits: 45,
-            property_access_cache_lookups: 50,
-            interned_types_count: 0,
-            interner_kb: 0.0,
-            has_query_cache: false,
-            has_def_store: false,
-            has_residency: false,
-            has_module_deps: false,
-            perf_counter_dump: String::new(),
-            ..DiagnosticsReport::default()
-        };
-        let out = render_diagnostics_report(&report, true);
-        let expected = "\n\
-Files:                         2\n\
-Lines of Library:              500\n\
-Lines of Definitions:          0\n\
-Lines of TypeScript:           100\n\
-Lines of JavaScript:           0\n\
-Lines of JSON:                 0\n\
-Lines of Other:                0\n\
-Errors:                        0\n\
-Total time:                    1.00s\n\
-Emitted files:                 1\n\
-Total diagnostics:             0\n\
-Request cache hits:            90\n\
-Request cache misses:          10\n\
-Request cache hit rate:        90.0%\n\
-Contextual cache bypasses:     2\n\
-clear_type_cache_recursive:    1\n\
-Access request-cache hit rate: 90.0% (45/50)\n\
-Memory used:                   8192K\n";
-        assert_eq!(out, expected, "extended golden mismatch:\n{out}");
-    }
-}
-
 fn reject_tsconfig_only_cli_options(args: &CliArgs) {
     for (name, values) in [
         ("paths", args.paths.as_ref()),
@@ -2346,57 +594,6 @@ fn reject_build_only_cli_options(args: &CliArgs) {
             std::process::exit(EXIT_DIAGNOSTICS_OUTPUTS_SKIPPED);
         }
     }
-}
-
-/// Get peak memory usage (max RSS) in KB.
-///
-/// Supported platforms:
-/// - **Unix (Linux + macOS)**: calls `getrusage(RUSAGE_SELF)` to read `ru_maxrss`.
-///   On macOS `ru_maxrss` is in bytes, on Linux it is in KB.
-/// - **Other**: returns 0.
-///
-/// This reports *peak* RSS (matching tsc's `--extendedDiagnostics` behavior).
-#[cfg(unix)]
-#[allow(unsafe_code)]
-fn get_memory_usage_kb() -> u64 {
-    // Minimal repr(C) struct matching the POSIX rusage layout.
-    // We only need fields through ru_maxrss; remaining fields are padding.
-    #[repr(C)]
-    struct Rusage {
-        ru_utime: [i64; 2], // struct timeval (tv_sec + tv_usec), 16 bytes on 64-bit
-        ru_stime: [i64; 2], // struct timeval
-        ru_maxrss: i64,     // max resident set size
-        _pad: [i64; 13],    // remaining fields (ixrss through nivcsw)
-    }
-
-    const RUSAGE_SELF: i32 = 0;
-
-    unsafe extern "C" {
-        fn getrusage(who: i32, usage: *mut Rusage) -> i32;
-    }
-
-    unsafe {
-        let mut usage: Rusage = std::mem::zeroed();
-        if getrusage(RUSAGE_SELF, &mut usage) == 0 {
-            let maxrss = usage.ru_maxrss as u64;
-            // macOS reports bytes; Linux reports KB.
-            #[cfg(target_os = "macos")]
-            {
-                maxrss / 1024
-            }
-            #[cfg(not(target_os = "macos"))]
-            {
-                maxrss
-            }
-        } else {
-            0
-        }
-    }
-}
-
-#[cfg(not(unix))]
-fn get_memory_usage_kb() -> u64 {
-    0
 }
 
 fn handle_init(args: &CliArgs, cwd: &std::path::Path) -> Result<()> {
@@ -3560,6 +1757,18 @@ fn handle_build_single_project(
 
 #[path = "tsz/show_config.rs"]
 mod show_config;
+
+#[path = "tsz/arg_preprocess.rs"]
+mod arg_preprocess;
+
+#[path = "tsz/clap_errors.rs"]
+mod clap_errors;
+
+#[path = "tsz/diagnostics_report.rs"]
+mod diagnostics_report;
+
+#[cfg(test)]
+use arg_preprocess::split_response_line;
 
 #[cfg(test)]
 #[path = "tsz/tests.rs"]
